fix: remove cluster entity
